Added dice roller methods
