+++ CharMaker/src/Race.java
@@ -3,4 +3,5 @@ import java.util.List;
 public class Race {
 	String name;
 	List<AbilityMod> abilityMods;
+	Size size;
 }
